src-tauri/diarizer/local_server/requirements.txt
fastapi
uvicorn
numpy
soundfile
torch
pyannote.audio==3.*

